requirements.txt
python-telegram-bot==13.13
Telethon==1.24.0
streamlit==1.13.0